requirements.txt
streamlit>=1.36.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.18
google-auth-oauthlib>=1.1.0
sqlalchemy>=2.0
python-dotenv>=1.0.0
pytest>=8.0.0
